refactor(streams): split tee() into streams/tee.rs (file-size gate)
streams.rs crossed the 2000-line lint gate with the lazy-tee implementation;
move the tee registries, helpers, and js_readable_stream_tee into a submodule
(same split as streams/expando.rs). No behavior change; the gap test is
unchanged and still byte-matches node.
